requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
fairlearn==0.12.0
gunicorn==23.0.0
joblib==1.4.2
mysqlclient==2.2.7
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.3.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
